refactor: Relocate notification logic to handlers
- integrate `handleNewNotification` to handle passing friend request data in `send friend request` and `friend request accepted` callbacks in `handleSocial.js`.
- integrate `handleNewNotification` to handle passing exam feedback user data in `exam finished` callbacks in `socket.exam.js`.
- remove `new notification` event emission from `handleFinishExam`, `handleIncomingRequest` and `handleAcceptRequest` functions in `useGlobalEvents.js`.
- remove redundant `receiverId` evaluation based on notification name in `handleNewNotification` in `socket.notification.js`.

Note: This change decouples notification generation from client-side socket listeners by moving the logic into the backend event handlers that trigger notifications.
